# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MLSystem2 Training UI API",
    "description": "Public FastAPI contract for the MLSystem2 training site.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "summary": "Login",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Login Api V1 Auth Login Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/mlflow/experiments": {
      "get": {
        "summary": "Get Mlflow Experiments",
        "operationId": "get_mlflow_experiments_api_v1_mlflow_experiments_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MLflowExperimentInfo"
                  },
                  "type": "array",
                  "title": "Response Get Mlflow Experiments Api V1 Mlflow Experiments Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Post Mlflow Experiment",
        "operationId": "post_mlflow_experiment_api_v1_mlflow_experiments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MLflowExperimentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MLflowExperimentInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/custom-datasets": {
      "post": {
        "summary": "Post Custom Dataset",
        "operationId": "post_custom_dataset_api_v1_custom_datasets_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_post_custom_dataset_api_v1_custom_datasets_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomDatasetInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/model-export/triton-zip": {
      "post": {
        "summary": "Post Model Export Triton Zip",
        "operationId": "post_model_export_triton_zip_api_v1_model_export_triton_zip_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_post_model_export_triton_zip_api_v1_model_export_triton_zip_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/results/training/{result_id}/triton-zip": {
      "post": {
        "summary": "Post Training Result Triton Zip",
        "operationId": "post_training_result_triton_zip_api_v1_results_training__result_id__triton_zip_post",
        "parameters": [
          {
            "name": "result_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Result Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_post_training_result_triton_zip_api_v1_results_training__result_id__triton_zip_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/automation/enabled": {
      "put": {
        "summary": "Put Automation Enabled",
        "operationId": "put_automation_enabled_api_v1_automation_enabled_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutomationEnabledUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutomationSnapshot"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/automation/rules": {
      "put": {
        "summary": "Put Automation Rule",
        "operationId": "put_automation_rule_api_v1_automation_rules_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutomationRuleUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutomationRuleInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/files/{file_id}/download": {
      "get": {
        "summary": "Download File",
        "operationId": "download_file_api_v1_files__file_id__download_get",
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "File Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AutomationEnabledUpdate": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "enabled"
        ],
        "title": "AutomationEnabledUpdate"
      },
      "AutomationRuleInfo": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "dataset_key": {
            "type": "string",
            "title": "Dataset Key"
          },
          "architecture": {
            "type": "string",
            "title": "Architecture"
          },
          "training_enabled": {
            "type": "boolean",
            "title": "Training Enabled",
            "default": false
          },
          "pseudo_markup_enabled": {
            "type": "boolean",
            "title": "Pseudo Markup Enabled",
            "default": false
          },
          "dataset_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Version"
          },
          "training_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResultStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "pseudo_markup_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResultStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "current_training_result_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Training Result Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "dataset_key",
          "architecture"
        ],
        "title": "AutomationRuleInfo"
      },
      "AutomationRuleUpdate": {
        "properties": {
          "dataset_key": {
            "type": "string",
            "title": "Dataset Key"
          },
          "architecture": {
            "type": "string",
            "title": "Architecture"
          },
          "training_enabled": {
            "type": "boolean",
            "title": "Training Enabled"
          },
          "pseudo_markup_enabled": {
            "type": "boolean",
            "title": "Pseudo Markup Enabled"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "dataset_key",
          "architecture",
          "training_enabled",
          "pseudo_markup_enabled"
        ],
        "title": "AutomationRuleUpdate"
      },
      "AutomationSnapshot": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "datasets": {
            "items": {
              "$ref": "#/components/schemas/DatasetInfo"
            },
            "type": "array",
            "title": "Datasets"
          },
          "models": {
            "items": {
              "$ref": "#/components/schemas/ModelInfo"
            },
            "type": "array",
            "title": "Models"
          },
          "rules": {
            "items": {
              "$ref": "#/components/schemas/AutomationRuleInfo"
            },
            "type": "array",
            "title": "Rules"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "enabled",
          "datasets",
          "models",
          "rules"
        ],
        "title": "AutomationSnapshot"
      },
      "Body_post_custom_dataset_api_v1_custom_datasets_post": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "default": "Custom"
          },
          "scenes_txt": {
            "anyOf": [
              {
                "type": "string",
                "format": "binary"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenes Txt"
          },
          "annotation_geojson": {
            "anyOf": [
              {
                "type": "string",
                "format": "binary"
              },
              {
                "type": "null"
              }
            ],
            "title": "Annotation Geojson"
          }
        },
        "type": "object",
        "title": "Body_post_custom_dataset_api_v1_custom_datasets_post"
      },
      "Body_post_model_export_triton_zip_api_v1_model_export_triton_zip_post": {
        "properties": {
          "model_name": {
            "type": "string",
            "title": "Model Name",
            "default": ""
          },
          "sample_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Size"
          },
          "checkpoint": {
            "anyOf": [
              {
                "type": "string",
                "format": "binary"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkpoint"
          }
        },
        "type": "object",
        "title": "Body_post_model_export_triton_zip_api_v1_model_export_triton_zip_post"
      },
...
      "Body_post_training_result_triton_zip_api_v1_results_training__result_id__triton_zip_post": {
        "properties": {
          "model_name": {
            "type": "string",
            "title": "Model Name",
            "default": ""
          },
          "sample_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Size"
          }
        },
        "type": "object",
        "title": "Body_post_training_result_triton_zip_api_v1_results_training__result_id__triton_zip_post"
      },
...
      "CustomDatasetInfo": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "scenes_file": {
            "$ref": "#/components/schemas/StoredFileInfo"
          },
          "annotation_file": {
            "$ref": "#/components/schemas/StoredFileInfo"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "scenes_file",
          "annotation_file",
          "created_at"
        ],
        "title": "CustomDatasetInfo"
      },
      "DatasetInfo": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "class_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Class Key"
          },
          "class_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Class Name"
          },
          "variant_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variant Key"
          },
          "variant_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variant Name"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "is_custom": {
            "type": "boolean",
            "title": "Is Custom",
            "default": false
          },
          "scenes_file": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenes File"
          },
          "annotation_file": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Annotation File"
          },
          "hard_negative_annotation_file": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hard Negative Annotation File"
          },
          "image_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Count"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "diagnostics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Diagnostics"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "key",
          "name"
        ],
        "title": "DatasetInfo"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest"
      },
      "MLflowExperimentCreate": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "MLflowExperimentCreate"
      },
      "MLflowExperimentInfo": {
        "properties": {
          "experiment_id": {
            "type": "string",
            "title": "Experiment Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "experiment_id",
          "name"
        ],
        "title": "MLflowExperimentInfo"
      },
      "ModelInfo": {
        "properties": {
          "architecture": {
            "type": "string",
            "title": "Architecture"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "input_channels": {
            "type": "integer",
            "title": "Input Channels"
          },
          "output_channels": {
            "type": "integer",
            "title": "Output Channels"
          },
          "pretrained": {
            "type": "boolean",
            "title": "Pretrained"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "architecture",
          "display_name",
          "input_channels",
          "output_channels",
          "pretrained"
        ],
        "title": "ModelInfo"
      },
...
      "ResultStatus": {
        "type": "string",
        "enum": [
          "running",
          "ok",
          "error",
          "cancelled"
        ],
        "title": "ResultStatus"
      },
...
      "StoredFileInfo": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "kind": {
            "$ref": "#/components/schemas/StoredFileKind"
          },
          "original_name": {
            "type": "string",
            "title": "Original Name"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "download_url": {
            "type": "string",
            "title": "Download Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "kind",
          "original_name",
          "size_bytes",
          "created_at",
          "download_url"
        ],
        "title": "StoredFileInfo"
      },
      "StoredFileKind": {
        "type": "string",
        "enum": [
          "scenes_txt",
          "annotation_geojson",
          "pseudo_markup_geojson"
        ],
        "title": "StoredFileKind"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}